+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog: rich-codex
 
+## Unreleased
+
+- ✨ MDX support: `.mdx` files are now searched by default and config comments can be written as JSX comments (`{/* RICH-CODEX ... */}`), as MDX doesn't allow HTML comments
+
 ## Version 1.2.11 (2025-04-22)
 
 - 🐛 Fix validation error ([#55](https://github.com/ewels/rich-codex/pull/55))